Update to v2.0.0-alpha.3
